# src/client/generated/w3s-openapi.yaml
openapi: 3.0.3
info:
  title: W3S
  description: All W3S APIs
  version: '1.0'
servers:
  - url: https://api.circle.com/v1/w3s
security:
  - BearerAuth: []
...
paths:
...
  /developer/transactions/transfer:
    post:
      tags:
        - Transactions
      summary: Create a transfer transaction
      description: Initiates an on-chain digital asset transfer from a specified developer-controlled wallet.
      operationId: createDeveloperTransactionTransfer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTransferTransactionForDeveloperRequest'
        description: Create transaction for developer request
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/CreateTransferTransactionForDeveloperResponse'
          description: OK
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
        default:
          $ref: '#/components/responses/Default'
  /developer/transactions/{id}/accelerate:
    post:
      tags:
        - Transactions
      summary: Accelerate a transaction
      description: Accelerates a specified on-chain digital asset transfer from a developer-controlled wallet. Additional gas fees may be incurred.
      operationId: createDeveloperTransactionAccelerate
      parameters:
        - name: id
          description: Transaction ID
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccelerateTransactionForDeveloperRequest'
        description: Accelerate transaction for developer request
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/AccelerateTransactionForDeveloperResponse'
          description: Transaction accelerated
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
        default:
          $ref: '#/components/responses/Default'
  /developer/transactions/{id}/cancel:
    post:
      tags:
        - Transactions
      summary: Cancel a transaction
      description: Cancels a specified on-chain digital asset transfer from a developer-controlled wallet. Gas fees may still be incurred.
      operationId: createDeveloperTransactionCancel
      parameters:
        - name: id
          description: Transaction ID
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        description: Cancel transaction for developer request
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancelTransactionForDeveloperRequest'
      responses:
        '200':
          description: Transaction canceled
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/CancelTransactionForDeveloperResponse'
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
        default:
          $ref: '#/components/responses/Default'
  /developer/transactions/contractExecution:
    post:
      tags:
        - Transactions
      summary: Create a contract execution transaction
      description: Creates a transaction which executes a smart contract. ABI parameters must be passed in the request.
      operationId: createDeveloperTransactionContractExecution
      requestBody:
        required: true
        description: Create transaction for developer request
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateContractExecutionTransactionForDeveloperRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/CreateContractExecutionTransactionForDeveloperResponse'
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
        default:
          $ref: '#/components/responses/Default'
  /transactions:
    get:
      tags:
        - Transactions
      summary: List transactions
      description: Lists all transactions. Includes details such as status, source/destination, and transaction hash.
      operationId: listTransactions
      parameters:
        - $ref: '#/components/parameters/XUserToken'
        - $ref: '#/components/parameters/Blockchain'
        - $ref: '#/components/parameters/CustodyType'
        - $ref: '#/components/parameters/DestinationAddress'
        - $ref: '#/components/parameters/IncludeAll'
        - $ref: '#/components/parameters/Operation'
        - $ref: '#/components/parameters/State'
        - $ref: '#/components/parameters/TxHash'
        - $ref: '#/components/parameters/TxType'
        - $ref: '#/components/parameters/WalletIds'
        - $ref: '#/components/parameters/From'
        - $ref: '#/components/parameters/To'
        - $ref: '#/components/parameters/PageBefore'
        - $ref: '#/components/parameters/PageAfter'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/GetTransactionsResponse'
          description: Transactions retrieved
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
            Link:
              description: 'Pagination cursor information. format: <https://api.circle.com/v1/developer/walletSets?to=2020-06-02T12:01:00Z&pageSize=50>; rel=\"next\'
              schema:
                type: string
        default:
          $ref: '#/components/responses/Default'
  /transactions/{id}:
    get:
      tags:
        - Transactions
      summary: Get a transaction
      description: Retrieves info for a single transaction using it's unique identifier.
      operationId: getTransaction
      parameters:
        - name: id
          description: Transaction id
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/XUserToken'
        - $ref: '#/components/parameters/TxType'
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/GetTransactionByIDResponse'
          description: Transaction retrieved
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
        default:
          $ref: '#/components/responses/Default'
  /transactions/transfer/estimateFee:
    post:
      tags:
        - Transactions
      summary: Estimate fee for a transfer transaction
      description: Estimates gas fees that will be incurred for a transfer transaction; given its amount, blockchain, and token.
      operationId: createTransferEstimateFee
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EstimateTransferTransactionFeeRequest'
        description: Estimate transaction fee request
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/EstimateTransactionFeeResponse'
          description: Transaction fee estimated
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
        default:
          $ref: '#/components/responses/Default'
  /transactions/contractExecution/estimateFee:
    post:
      tags:
        - Transactions
      summary: Estimate fee for a contract execution transaction
      description: Estimates gas fees that will be incurred for a contract execution transaction, given its ABI parameters and blockchain.
      operationId: createTransactionEstimateFee
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EstimateContractExecutionTransactionFeeRequest'
        description: Estimate transaction fee request
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/EstimateTransactionFeeResponse'
          description: Transaction fee estimated
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
        default:
          $ref: '#/components/responses/Default'
  /transactions/validateAddress:
    post:
      tags:
        - Transactions
      summary: Validate an address
      description: Confirms that a specified address is valid for a given token on a certain blockchain.
      operationId: createValidateAddress
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ValidateAddressRequest'
        description: Validate address request
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/ValidateAddressResponse'
          description: Address validated
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
        default:
          $ref: '#/components/responses/Default'
...
  /user/transactions/transfer:
    post:
      tags:
        - Transactions
      summary: Create a challenge for a transfer
      description: Generates a challenge for initiating an on-chain digital asset transfer from a specified user-controlled wallet
      operationId: createUserTransactionTransferChallenge
      parameters:
        - $ref: '#/components/parameters/XUserTokenRequired'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTransferTransactionForEndUserRequest'
        description: Create transaction for end user request
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/CreateTransferTransactionForEndUserResponse'
          description: OK
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
        default:
          $ref: '#/components/responses/Default'
  /user/transactions/{id}/accelerate:
    post:
      tags:
        - Transactions
      summary: Create a Challenge to accelerate a transaction
      description: Generates a challenge for accelerating a specified on-chain digital asset transfer from a user-controlled wallet. Additional gas fees may be incurred.
      operationId: createUserTransactionAccelerateChallenge
      parameters:
        - name: id
          description: Transaction ID
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/XUserTokenRequired'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccelerateTransactionForEndUserRequest'
        description: Accelerate transaction for end user request
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/AccelerateTransactionForEndUserResponse'
          description: Transaction accelerated
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
        default:
          $ref: '#/components/responses/Default'
  /user/transactions/{id}/cancel:
    post:
      tags:
        - Transactions
      summary: Create a challenge to cancel a transaction
      description: Generates a challenge for cancelling a specified on-chain digital asset transfer from a user-controlled wallet. Gas fees may still be incurred.
      operationId: createUserTransactionCancelChallenge
      parameters:
        - name: id
          description: Transaction ID
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/XUserTokenRequired'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancelTransactionForEndUserRequest'
        description: Cancel transaction for end user request
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/CancelTransactionForEndUserResponse'
          description: Transaction cancelled
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
        default:
          $ref: '#/components/responses/Default'
  /user/transactions/contractExecution:
    post:
      tags:
        - Transactions
      summary: Create a challenge for contract execution
      description: Generates a challenge for creating a transaction which executes a smart contract. ABI parameters must be passed in the request.
      operationId: createUserTransactionContractExecutionChallenge
      parameters:
        - $ref: '#/components/parameters/XUserTokenRequired'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContractExecutionTxnPinCodeRequest'
        description: Create transaction for end user request
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/CreateContractExecutionTransactionForEndUserResponse'
          description: OK
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
        default:
          $ref: '#/components/responses/Default'
...
components:
  schemas:
...
    CreateTransferTransactionForDeveloperRequest:
      type: object
      required:
        - destinationAddress
        - entitySecretCiphertext
        - idempotencyKey
        - tokenId
        - walletId
      properties:
        idempotencyKey:
          $ref: '#/components/schemas/schemas_IdempotencyKey'
        amounts:
          $ref: '#/components/schemas/TransferAmounts'
        destinationAddress:
          $ref: '#/components/schemas/DestinationAddress'
        entitySecretCiphertext:
          $ref: '#/components/schemas/EntitySecretCiphertext'
        feeLevel:
          $ref: '#/components/schemas/FeeLevel'
        gasLimit:
          $ref: '#/components/schemas/GasLimit'
        gasPrice:
          $ref: '#/components/schemas/GasPrice'
        maxFee:
          $ref: '#/components/schemas/MaxFee'
        priorityFee:
          $ref: '#/components/schemas/PriorityFee'
        nftTokenIds:
          type: array
          items:
            type: string
          description: NftTokenIDs specify the nft ID list to be transferred/batchTransferred for NFT withdrawal. Note. Only erc1155 supports safeBatchTransferFrom.
        refId:
          $ref: '#/components/schemas/TransactionReferenceId'
        tokenId:
          description: The unique system generated identifier associated with the token.
          example: a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11
          format: uuid
          type: string
        walletId:
          $ref: '#/components/schemas/schemas_WalletId'
...
    CreateTransferTransactionForDeveloperResponse:
      type: object
      required:
        - id
        - state
      properties:
        id:
          $ref: '#/components/schemas/schemas_Id'
        state:
          $ref: '#/components/schemas/TransactionState'
    AccelerateTransactionForDeveloperRequest:
      type: object
      required:
        - entitySecretCiphertext
        - idempotencyKey
      properties:
        idempotencyKey:
          $ref: '#/components/schemas/schemas_IdempotencyKey'
        entitySecretCiphertext:
          $ref: '#/components/schemas/EntitySecretCiphertext'
    AccelerateTransactionForDeveloperResponse:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/schemas_Id'
    CancelTransactionForDeveloperRequest:
      type: object
      required:
        - entitySecretCiphertext
        - idempotencyKey
      properties:
        idempotencyKey:
          $ref: '#/components/schemas/schemas_IdempotencyKey'
        entitySecretCiphertext:
          $ref: '#/components/schemas/EntitySecretCiphertext'
    CancelTransactionForDeveloperResponse:
      type: object
      required:
        - id
        - state
      properties:
        id:
          $ref: '#/components/schemas/schemas_Id'
        state:
          $ref: '#/components/schemas/TransactionState'
...
    CreateContractExecutionTransactionForDeveloperRequest:
      type: object
      required:
        - abiParameters
        - contractAddress
        - entitySecretCiphertext
        - idempotencyKey
        - walletId
      properties:
        idempotencyKey:
          $ref: '#/components/schemas/schemas_IdempotencyKey'
        abiFunctionSignature:
          $ref: '#/components/schemas/schemas_AbiFunctionSignature'
        abiParameters:
          $ref: '#/components/schemas/schemas_AbiParameters'
        contractAddress:
          $ref: '#/components/schemas/ContractAddress'
        entitySecretCiphertext:
          $ref: '#/components/schemas/EntitySecretCiphertext'
        feeLevel:
          $ref: '#/components/schemas/FeeLevel'
        gasLimit:
          $ref: '#/components/schemas/GasLimit'
        gasPrice:
          $ref: '#/components/schemas/GasPrice'
        maxFee:
          $ref: '#/components/schemas/MaxFee'
        priorityFee:
          $ref: '#/components/schemas/PriorityFee'
        refId:
          $ref: '#/components/schemas/TransactionReferenceId'
        walletId:
          $ref: '#/components/schemas/schemas_WalletId'
    CreateContractExecutionTransactionForDeveloperResponse:
      type: object
      required:
        - id
        - state
      properties:
        id:
          description: Unique system generated identifier of the transaction.
          type: string
        state:
          $ref: '#/components/schemas/TransactionState'
...
    GetTransactionsResponse:
      type: object
      properties:
        transactions:
          type: array
          items:
            $ref: '#/components/schemas/Transaction'
    GetTransactionByIDResponse:
      properties:
        transaction:
          $ref: '#/components/schemas/Transaction'
      type: object
...
    EstimateTransferTransactionFeeRequest:
      type: object
      required:
        - amounts
        - destinationAddress
        - tokenId
      properties:
        amounts:
          $ref: '#/components/schemas/TransferAmounts'
        destinationAddress:
          $ref: '#/components/schemas/DestinationAddress'
        nftTokenIds:
          $ref: '#/components/schemas/NftTokenIds'
        sourceAddress:
          $ref: '#/components/schemas/schemas_SourceAddress'
        tokenId:
          $ref: '#/components/schemas/TokenId'
        walletId:
          $ref: '#/components/schemas/schemas_WalletId'
    EstimateTransactionFeeResponse:
      type: object
      properties:
        high:
          $ref: '#/components/schemas/EstimatedTransactionFee'
        low:
          $ref: '#/components/schemas/EstimatedTransactionFee'
        medium:
          $ref: '#/components/schemas/EstimatedTransactionFee'
    EstimateContractExecutionTransactionFeeRequest:
      type: object
      required:
        - abiFunctionSignature
        - abiParameters
        - contractAddress
      properties:
        abiFunctionSignature:
          $ref: '#/components/schemas/schemas_AbiFunctionSignature'
        abiParameters:
          $ref: '#/components/schemas/schemas_AbiParameters'
        blockchain:
          description: |
            Blockchain associated with the transaction. Required along with sourceAddress
            if walletId is not provided. The blockchain and walletId fields are mutually exclusive.
          allOf:
            - $ref: '#/components/schemas/Blockchain'
        contractAddress:
          $ref: '#/components/schemas/ContractAddress'
        sourceAddress:
          allOf:
            - $ref: '#/components/schemas/schemas_SourceAddress'
          description: |
            Source address of the transaction. Required along with blockchain if walletId is
            not provided. The sourceAddress and walletId fields are mutually exclusive.
        walletId:
          allOf:
            - $ref: '#/components/schemas/schemas_WalletId'
          description: |
            The source wallet id of the transaction. Required if sourceAddress and blockchain are not provided. The walletId field is mutually exclusive to the blockchain and sourceAddress fields.
    ValidateAddressRequest:
      type: object
      required:
        - address
        - blockchain
      properties:
        address:
          $ref: '#/components/schemas/BlockchainAddress'
        blockchain:
          $ref: '#/components/schemas/Blockchain'
    ValidateAddressResponse:
      type: object
      required:
        - isValid
      properties:
        isValid:
          type: boolean
...
    CreateTransferTransactionForEndUserRequest:
      type: object
      required:
        - destinationAddress
        - idempotencyKey
        - tokenId
        - walletId
      properties:
        idempotencyKey:
          $ref: '#/components/schemas/schemas_IdempotencyKey'
        amounts:
          $ref: '#/components/schemas/TransferAmounts'
        destinationAddress:
          $ref: '#/components/schemas/DestinationAddress'
        feeLevel:
          $ref: '#/components/schemas/FeeLevel'
        gasLimit:
          $ref: '#/components/schemas/GasLimit'
        gasPrice:
          $ref: '#/components/schemas/GasPrice'
        maxFee:
          $ref: '#/components/schemas/MaxFee'
        priorityFee:
          $ref: '#/components/schemas/PriorityFee'
        nftTokenIds:
          $ref: '#/components/schemas/NftTokenIds'
        refId:
          $ref: '#/components/schemas/TransactionReferenceId'
        tokenId:
          $ref: '#/components/schemas/TokenId'
        walletId:
          $ref: '#/components/schemas/schemas_WalletId'
    CreateTransferTransactionForEndUserResponse:
      type: object
      required:
        - challengeId
      properties:
        challengeId:
          $ref: '#/components/schemas/ChallengeId'
    AccelerateTransactionForEndUserRequest:
      required:
        - idempotencyKey
      type: object
      properties:
        idempotencyKey:
          $ref: '#/components/schemas/schemas_IdempotencyKey'
    AccelerateTransactionForEndUserResponse:
      type: object
      properties:
        challengeId:
          $ref: '#/components/schemas/ChallengeId'
    CancelTransactionForEndUserRequest:
      type: object
      required:
        - idempotencyKey
      properties:
        idempotencyKey:
          $ref: '#/components/schemas/schemas_IdempotencyKey'
    CancelTransactionForEndUserResponse:
      type: object
      required:
        - challengeId
      properties:
        challengeId:
          $ref: '#/components/schemas/ChallengeId'
    ContractExecutionTxnPinCodeRequest:
      type: object
      required:
        - contractAddress
        - idempotencyKey
        - walletId
      properties:
        idempotencyKey:
          $ref: '#/components/schemas/schemas_IdempotencyKey'
        abiFunctionSignature:
          $ref: '#/components/schemas/schemas_AbiFunctionSignature'
        abiParameters:
          $ref: '#/components/schemas/schemas_AbiParameters'
        contractAddress:
          $ref: '#/components/schemas/ContractAddress'
        feeLevel:
          $ref: '#/components/schemas/FeeLevel'
        gasLimit:
          $ref: '#/components/schemas/GasLimit'
        gasPrice:
          $ref: '#/components/schemas/GasPrice'
        maxFee:
          $ref: '#/components/schemas/MaxFee'
        priorityFee:
          $ref: '#/components/schemas/PriorityFee'
        refId:
          $ref: '#/components/schemas/TransactionReferenceId'
        walletId:
          $ref: '#/components/schemas/schemas_WalletId'
    CreateContractExecutionTransactionForEndUserResponse:
      type: object
      required:
        - challengeId
      properties:
        challengeId:
          $ref: '#/components/schemas/ChallengeId'
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: PREFIX:ID:SECRET
      description: Circle's API Keys are formatted in the following structure "PREFIX:ID:SECRET". All three parts are requred to make a successful request.
  headers:
    XRequestId:
      description: Universally unique identifier (UUID v4) for the request. Helpful for identifying a request when communicating with Circle support.
      schema:
        type: string
        format: uuid
        example: 2adba88e-9d63-44bc-b975-9b6ae3440dde
  parameters:
    From:
      name: from
      description: From is the start time of the query, inclusive
      in: query
      schema:
        type: string
        format: date-time
        example: '2023-01-01T12:04:05Z'
    To:
      name: to
      description: To is the end time of the query, inclusive, default to current time
      in: query
      schema:
        type: string
        format: date-time
        example: '2023-01-01T12:04:05Z'
    PageBefore:
      name: pageBefore
      description: PageBefore is used to return items before the given item exclusively. SHOULD NOT be used in conjunction with pageAfter
      in: query
      schema:
        type: string
        format: uuid
    PageAfter:
      name: pageAfter
      description: PageAfter is used to return items after the given item exclusively. SHOULD NOT be used in conjunction with pageBefore
      in: query
      schema:
        type: string
        format: uuid
    PageSize:
      name: pageSize
      description: PageSize is the number of items to return
      in: query
      schema:
        type: integer
        default: 10
        maximum: 50
        minimum: 1
    Blockchain:
      name: blockchain
      description: Blockchain filters on the blockchain of the transaction
      in: query
      schema:
        enum:
          - ETH
          - ETH-GOERLI
          - AVAX
          - AVAX-FUJI
          - MATIC
          - MATIC-MUMBAI
        type: string
...
    XUserToken:
      name: X-User-Token
      description: User token
      in: header
      schema:
        type: string
...
    IncludeAll:
      name: includeAll
      in: query
      required: false
      schema:
        type: boolean
        example: true
...
    CustodyType:
      name: custodyType
      description: CustodyType filters on the CustodyType of the transaction
      in: query
      schema:
        enum:
          - DEVELOPER
          - ENDUSER
        type: string
    DestinationAddress:
      name: destinationAddress
      description: DestinationAddress filters on the DestinationAddress of the transaction
      in: query
      schema:
        type: string
    Operation:
      name: operation
      description: Operation filters on the Operation of the transaction
      in: query
      schema:
        enum:
          - TRANSFER
          - CONTRACT_EXECUTION
          - CONTRACT_DEPLOYMENT
        type: string
    State:
      name: state
      description: State filters on the state of the transaction
      in: query
      schema:
        type: string
        enum:
          - INITIATED
          - PENDING_RISK_SCREENING
          - DENIED
          - QUEUED
          - SENT
          - CONFIRMED
          - COMPLETED
          - CANCELLED
          - FAILED
    TxHash:
      name: txHash
      description: TxHash filters on the TxHash of the transaction
      in: query
      schema:
        type: string
    TxType:
      name: txType
      description: TxType filters on the TxType of the transaction
      in: query
      schema:
        enum:
          - INBOUND
          - OUTBOUND
        type: string
    WalletIds:
      name: walletIds
      description: WalletIDs filters on the wallet ids of the transaction
      in: query
      schema:
        type: string
...
    XUserTokenRequired:
      name: X-User-Token
      description: User token
      required: true
      in: header
      schema:
        type: string
...
  responses:
    Default:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
      description: Error response
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'